[CHORE] MainView -> ShareView 연결
